Labs/LAB-1/pyproject.toml
[project]
name = "lab-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker>=37.6.0",
    "matplotlib>=3.10.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
